# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * House Floor Plan Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Plot facing direction
 */
export type FloorPlanInputFacing = typeof FloorPlanInputFacing[keyof typeof FloorPlanInputFacing];


export const FloorPlanInputFacing = {
  North: 'North',
  South: 'South',
  East: 'East',
  West: 'West',
} as const;

export interface FloorPlanInput {
  /** Plot width in feet */
  plotWidth: number;
  /** Plot length in feet */
  plotLength: number;
  /**
     * Number of floors
     * @minimum 1
     * @maximum 4
     */
  floors: number;
  /** Plot facing direction */
  facing: FloorPlanInputFacing;
  /**
     * @minimum 1
     * @maximum 10
     */
  bedrooms: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  bathrooms: number;
  hasParking?: boolean;
  hasGarden?: boolean;
  hasPooja?: boolean;
  hasStudyRoom?: boolean;
  vastuCompliant?: boolean;
  /** @nullable */
  additionalNotes?: string | null;
}

export interface Room {
  name: string;
  /** X position in feet from origin */
  x: number;
  /** Y position in feet from origin */
  y: number;
  /** Room width in feet */
  width: number;
  /** Room length in feet */
  length: number;
  /** Floor number (0 = ground, 1 = first, etc.) */
  floor: number;
  /** @nullable */
  color?: string | null;
}

export interface FloorPlanResult {
  rooms: Room[];
  plotWidth: number;
  plotLength: number;
  floors: number;
  /** AI-generated description of the floor plan */
  description: string;
  /** @nullable */
  vastuNotes?: string | null;
}