# lib/api-zod/src/generated/types/floorPlanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * House Floor Plan Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { FloorPlanInputFacing } from './floorPlanInputFacing';

export interface FloorPlanInput {
  /** Plot width in feet */
  plotWidth: number;
  /** Plot length in feet */
  plotLength: number;
  /**
     * Number of floors
     * @minimum 1
     * @maximum 4
     */
  floors: number;
  /** Plot facing direction */
  facing: FloorPlanInputFacing;
  /**
     * @minimum 1
     * @maximum 10
     */
  bedrooms: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  bathrooms: number;
  hasParking?: boolean;
  hasGarden?: boolean;
  hasPooja?: boolean;
  hasStudyRoom?: boolean;
  vastuCompliant?: boolean;
  /** @nullable */
  additionalNotes?: string | null;
}