# lib/api-zod/src/generated/types/floorPlanResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * House Floor Plan Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { Room } from './room';

export interface FloorPlanResult {
  rooms: Room[];
  plotWidth: number;
  plotLength: number;
  floors: number;
  /** AI-generated description of the floor plan */
  description: string;
  /** @nullable */
  vastuNotes?: string | null;
}