# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * House Floor Plan Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  name: string;
  /** X position in feet from origin */
  x: number;
  /** Y position in feet from origin */
  y: number;
  /** Room width in feet */
  width: number;
  /** Room length in feet */
  length: number;
  /** Floor number (0 = ground, 1 = first, etc.) */
  floor: number;
  /** @nullable */
  color?: string | null;
}